Отменяет изменения.
